+++ plugin.audio.deutschlandfunk/addon.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<addon id="plugin.audio.deutschlandfunk" name="Deutschlandfunk" version="2.0.0" provider-name="Heckie">
+<addon id="plugin.audio.deutschlandfunk" name="Deutschlandfunk" version="2.0.1" provider-name="Heckie">
     <requires>
         <import addon="xbmc.python" version="3.0.0" />
         <import addon="script.module.requests" version="2.25.1" />
@@ -25,6 +25,10 @@ sämtliche Podcasts der Sender gestreamt werden.</description>
         <website>https://github.com/Heckie75/kodi-addon-deutschlandfunk</website>
         <source>https://github.com/Heckie75/kodi-addon-deutschlandfunk/tree/master/plugin.audio.deutschlandfunk</source>
         <news>
+v2.0.1 (2021-06-27)
+- Fixed incorrect navigation from specific podcast to station's podcasts menu when using Kodi Kore App 
+- Refactoring
+
 v2.0.0 (2021-05-13)
 - Migration to Kodi 19 (Matrix)
 - almost total rewrite
